add the generation finetune
